# backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "PROJECT NAME APIs",
      "version": "1.0.0",
      "description": "API documentation for Your App"
    },
    "tags": [
      {
        "name": "authentication-controller",
        "description": "Operations related to authentication"
      }
    ],
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
    "paths": {
      "/auth/signup": {
        "post": {
          "summary": "Save user",
          "tags": ["authentication-controller"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "User saved successfully",
              "content": {
                "application/json": {}
              }
            }
          }
        }
      },
      "/auth/signin": {
        "post": {
          "summary": "Signin",
          "tags": ["authentication-controller"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "required": true
                    },
                    "password": {
                      "email" : {
                        "type" : {
                          "type" : "string",
                          "required" : true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "User signin successful",
              "content": {
                "application/json": {}
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "User": {
          "type": "object",
          "properties": {
            "phone": {
              "type": "string",
              "required": true
            },
            "fullName": {
              "type": "string",
              "required": true
            },
            "email": {
              "type": "string",
              "required": true
            },
            "nationalId" : {
              "type": "string",
              "required": true 
            },
            "password": {
              "type": "string",
              "required": true
            },
            "role": {
              "type": "string",
              "required": true,
              "enum": ["ADMIN"]
            }
          }
        }
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }, 
    "security" : [
      {
        "bearerAuth" : []
      }
    ]
  }
  